\name{VoroGaze-package}
\alias{VoroGaze-package}
\alias{VoroGaze}
\docType{package}
\title{VoroGaze Research Workbench}
\description{
Run a local workbench for mapping fixation reports, checking screen and image
geometry, defining facial areas of interest, and exporting annotated or
participant-level results.
}
